# api/openapi-aws.yaml
openapi: 3.0.3
info:
  title: FlexQ API (AWS)
  version: 1.0.0
  description: |
    This API serves mock data for FlexQ application.
    It is designed to be used with AWS API Gateway + Lambda.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://YOUR_AWS_ACCOUNT_ID.execute-api.ap-northeast-1.amazonaws.com/dev
paths:
...
  /data/auth/login:
    post:
      summary: Invoke post-auth-login
      operationId: post_auth_login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              examples:
                AUTH_DATA:
                  $ref: '#/components/examples/AUTH_DATA'
        '400':
          description: Bad Request
          content:
            application/json:
              example:
                message: error
      x-amazon-apigateway-integration:
        type: aws_proxy
        httpMethod: POST
        uri: arn:aws:apigateway:ap-northeast-1:lambda:path/2015-03-31/functions/arn:aws:lambda:ap-northeast-1:YOUR_AWS_ACCOUNT_ID:function:post-auth-login/invocations
  /data/auth/register:
    post:
      summary: Invoke post-auth-register
      operationId: post_auth_register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - email
                - password
              properties:
                username:
                  type: string
                email:
                  type: string
                password:
                  type: string
      responses:
        '201':
          description: Created
        '400':
          description: Bad Request
        '409':
          description: Email already in use
      x-amazon-apigateway-integration:
        type: aws_proxy
        httpMethod: POST
        uri: arn:aws:apigateway:ap-northeast-1:lambda:path/2015-03-31/functions/arn:aws:lambda:ap-northeast-1:YOUR_AWS_ACCOUNT_ID:function:post-auth-register/invocations
  /data/auth/google:
    post:
      summary: Invoke post-auth-google
      operationId: post_auth_google
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - accessToken
              properties:
                accessToken:
                  type: string
      responses:
        '200':
          description: OK (existing user signed in)
        '201':
          description: Created (new user registered and signed in)
        '400':
          description: Bad Request
        '401':
          description: Invalid Google access token
      x-amazon-apigateway-integration:
        type: aws_proxy
        httpMethod: POST
        uri: arn:aws:apigateway:ap-northeast-1:lambda:path/2015-03-31/functions/arn:aws:lambda:ap-northeast-1:YOUR_AWS_ACCOUNT_ID:function:post-auth-google/invocations
  /data/auth/reset-password:
    post:
      summary: Invoke post-auth-reset-password
      operationId: post_auth_reset_password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - newPassword
              properties:
                email:
                  type: string
                newPassword:
                  type: string
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '404':
          description: User not found
      x-amazon-apigateway-integration:
        type: aws_proxy
        httpMethod: POST
        uri: arn:aws:apigateway:ap-northeast-1:lambda:path/2015-03-31/functions/arn:aws:lambda:ap-northeast-1:YOUR_AWS_ACCOUNT_ID:function:post-auth-reset-password/invocations
components:
  examples:
    AUTH_DATA:
      value: |-
        [
          {
            "id": "user_001",
            "email": "testuser@example.com",
            "password": "password123",
            "username": "User Profile Name",
            "thumbnail": "http://localhost:3000/images/sample/profile.jpg",
            "socialAccounts": [
              {
                "provider": "x",
                "username": "@user_name",
                "isLinked": true
              },
              {
                "provider": "instagram",
                "username": "@user_name",
                "isLinked": true
              },
              {
                "provider": "google",
                "username": "Yoshi Watanabe",
                "isLinked": true
              }
            ],
            "token": {
              "accessToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.mock_access_token",
              "refreshToken": "mock_refresh_token_value",
              "expiresIn": 3600
            }
          }
        ]
...